Packer: add checks to FMS Vagrant build scripts
Sometimes network operations in these scripts don't complete,
but Packer weirdly skips onwards anyway without failing. This
leads to builds without SSH keys or VirtualBox Guest Additions
which can be confusing. These changes attempt to work around this.

+++ packer/scripts/01_vagrant.sh
@@ -5,6 +5,7 @@ echo "==> Setting up Vagrant user."
 mkdir -p /home/vagrant/.ssh
 chmod 700 /home/vagrant/.ssh
 curl -L -o /home/vagrant/.ssh/authorized_keys https://raw.githubusercontent.com/mitchellh/vagrant/master/keys/vagrant.pub
+[ -e /home/vagrant/.ssh/authorized_keys ] || { echo "==> failed to download insecure key."; exit 1; }
 chown -R vagrant:vagrant /home/vagrant/.ssh
 chmod 0600 /home/vagrant/.ssh/*
 echo 'vagrant ALL=(ALL) NOPASSWD:ALL' > /etc/sudoers.d/vagrant
